Issue #4 - Title finder
Managed to write a function to find title by searching for the text with biggest size/font.

The commented block is for finding authors. FYI, will only work if the name is a line under the keyword Authors. Because of that, it is placed lower in the priority list since the difficulty is too high.